php/opcache-exporter: initial commit
This initial commit includes a proof-of-concept to call FastCGI to execute a
script using tokens to check and deny unauthorized access. It is built on
the php/fastcgi-client example.

This proof-of-concept is intended to save progress on an opcache exporter idea,
and will be replaced by a ReactPHP server in a future changeset